+++ www/reference/gatsby-node.js
@@ -331,7 +331,7 @@ const createAllPages = (sections, api, siteData, createPage, template) => {
         api: api,
         title: edge.section.section_name,
         description: edge.section.schema ? edge.section.schema.description : "",
-        to: { section: baseURL, method: null },
+        to: { section: baseURL, method: null, sectionObj: edge.section },
       },
     })
     edge.section.paths.forEach(p => {
@@ -345,7 +345,7 @@ const createAllPages = (sections, api, siteData, createPage, template) => {
             api: api,
             title: method.summary,
             description: method.description || "",
-            to: { section: baseURL, method: methodURL },
+            to: { section: baseURL, method: methodURL, sectionObj: edge.section },
           },
         })
       })