+++ www/reference/src/components/CopyToClipboard/index.js
@@ -1,9 +1,10 @@
+import { Box, Text } from "theme-ui"
 import React, { useState } from "react"
-import { Flex, Box, Text } from "theme-ui"
+
 import Clipboard from '../icons/clipboard'
 import ReactTooltip from "react-tooltip"
-import styled from "@emotion/styled"
 import copy from 'copy-to-clipboard'
+import styled from "@emotion/styled"
 
 const StyledTooltip = ({id, text}) => {
   const StyledTooltip = styled(ReactTooltip)`